turn outer map into sequential loop
this is because futhark does not parallelise bodies of sequential
loops if they are already within a parallel construct (the map).

we also make explicit the assumption that all instances are the same
size (i.e. regular input) although this assumption should be easy
to relax for the multi version.

todo: in non-multi versions, move the map/loop over the different instances
inside of the while-loop.